Port unsupported type validations to validations.rs
